Added powerups to the game
Change log:
1. Added powerups to the game
2. Added sprites for powerups
3. Minor bug fixes

+++ Space_Invaders/classes/Game/Difficulty.py
@@ -34,7 +34,7 @@ class Difficulty(object):
         if mul*value > 1:
 
             #Return the multiplier with the value
-            return mul*value
+            return int(mul*value)
 
         else:
